changed the feild type in class
Previously it had only two feild, for the representation of a line.but, it is
not enough. so,it is now an object of object that contains the information that
represents the line. object contains two key-value pair each represents
end of the line .
Accordingly , toString() method also changed, that represents both end.
And isEqual function is checking all ends of the line